LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_pointerlock_after_pointercapture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events pointer lock test

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: This test checks that we release the exsiting pointer capture when any element in the page gets a pointer lock.
Press left button down on the green rectangle and hold it.
Move the mouse inside the green rectangle.
Click right button while keeping left button down
Keep moving the mouse inside the green rectangle.
Test passes if the pointer capture is released on the yellow rectangle when the green rectangle gets the pointer lock.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN mouse no pointercapture while pointerlock
